requirements.txt
aniso8601==7.0.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
bcrypt==4.3.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.1.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
colorthief==0.2.1
commonmark==0.9.1
cryptography==43.0.3
dateparser==1.2.1
defusedxml==0.8.0rc2
distro==1.9.0
Django==5.1.7
django-admin-soft-dashboard==1.0.11
django-colorfield==0.12.0
django-filter==25.1
django-jazzmin==3.0.1
django-material==1.12.0
django-unfold==0.52.0
django-viewflow==2.2.9
djangorestframework==3.14.0
dnspython==2.7.0
flatbuffers==25.2.10
ggshield==1.37.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
jet-bridge-base==1.11.15
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.44.0
marshmallow==3.18.0
marshmallow-dataclass==8.5.14
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.60.2
opencv-python-headless==4.11.0.86
packaging==24.2
paramiko==3.5.1
pillow==11.1.0
platformdirs==3.0.0
pooch==1.8.2
promise==2.3
prompt-toolkit==2.0.9
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
psycopg2-binary==2.9.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pygitguardian==1.20.0
Pygments==2.19.1
PyJWT==2.6.0
PyMatting==1.1.13
pymongo==4.1.1
PyNaCl==1.5.0
pyparsing==3.2.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
rembg==2.0.64
requests==2.32.3
requests-oauthlib==2.0.0
rich==12.5.1
rpds-py==0.23.1
rsa==4.9
Rx==1.6.3
scikit-image==0.25.2
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
social-auth-core==4.4.2
SQLAlchemy==1.4.54
sqlparse==0.5.3
sshtunnel==0.4.0
sympy==1.13.3
tifffile==2025.2.18
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
whitenoise==6.5.0
